get by id, delete, update

+++ src/main/java/com/mimie/cetimBackend/repository/EquipmentRepository.java
@@ -3,5 +3,5 @@ package com.mimie.cetimBackend.repository;
 import com.mimie.cetimBackend.model.Equipment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EquipmentRepository extends JpaRepository<Equipment,Integer> {
+public interface EquipmentRepository extends JpaRepository<Equipment,Long> {
 }
